+++ test/unittests/test_id_to_iri.py
@@ -26,20 +26,20 @@ class TestIdToIri(unittest.TestCase):
     def test_invalid_xml_file_name(self) -> None:
         with self.assertRaisesRegex(BaseError, r"File test\.xml could not be found"):
             id_to_iri(xml_file='test.xml',
-                      json_file='testdata/test-id2iri-mapping.json',
+                      json_file='testdata/id2iri/test-id2iri-mapping.json',
                       out_file=self.out_file,
                       verbose=True)
 
     def test_invalid_json_file_name(self) -> None:
         with self.assertRaisesRegex(BaseError, r"File test\.json could not be found"):
-            id_to_iri(xml_file='testdata/test-id2iri-data.xml',
+            id_to_iri(xml_file='testdata/id2iri/test-id2iri-data.xml',
                       json_file='test.json',
                       out_file=self.out_file,
                       verbose=True)
 
     def test_replace_id_with_iri(self) -> None:
-        id_to_iri(xml_file='testdata/test-id2iri-data.xml',
-                  json_file='testdata/test-id2iri-mapping.json',
+        id_to_iri(xml_file='testdata/id2iri/test-id2iri-data.xml',
+                  json_file='testdata/id2iri/test-id2iri-mapping.json',
                   out_file=self.out_file,
                   verbose=True)